create bin dir
