+++ lib/html-manifest-stream.js
@@ -1,0 +1,11 @@
+var hyperstream = require('hyperstream')
+
+module.exports = manifestStream
+
+function manifestStream (manifest) {
+  return hyperstream({
+    head: {
+      _append: manifest
+    }
+  })
+}
