+++ package.json
@@ -36,6 +36,7 @@
     "disc": "^1.3.2",
     "envify": "^4.0.0",
     "explain-error": "^1.0.3",
+    "findup": "^0.1.5",
     "from2": "^2.3.0",
     "from2-string": "^1.1.0",
     "html-minifier": "^3.4.3",
@@ -51,6 +52,7 @@
     "resolve": "^1.1.7",
     "run-parallel": "^1.1.6",
     "send": "^0.15.0",
+    "server-router": "^5.2.0",
     "sheetify": "^6.0.1",
     "subarg": "^1.0.0",
     "temp-path": "^1.0.0",
